third_party/WebKit/LayoutTests/external/wpt/longtask-timing/longtask-in-parentiframe-expected.txt

This is a testharness.js-based test.
FAIL Performance longtask entries in parent are observable in child iframe assert_equals: expected "longtask+same-origin-ancestor+frame" but got "longtask+same-origin-ancestor+script"
Harness: the test ran to completion.
